Create workspace dir in server instead of workspace pod on Openshift
Create workspace directory on PVC before launching workspace instead
of starting a Pod to create the directory. This step is necessary
because if a workspace directory does not exist, it will be created
with root permissions when mounted on a subpath.

Assumes that the workspace PVC is mounted in the server pod, on path
/projects.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>